move group accumulator adapter to functions-aggregate-common
